+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Add `paymentReaderCardDetails` parameter to `Transaction.sale`
 * Add `skipAdvancedFraudChecking` to `Customer.create` and `Customer.update`
 * Add webhook sample for `GrantedPaymentMethodRevoked`